CodeGen.CXX.Code

Description

A writer monad for building up C++ code, to use Haskell as a (way more powerful) replacement for the C preprocessor. Unlike that, it is designed to produce reasonably human-readable code, so the resulting files can also be used as "proper" code files, without Haskell available.

Documentation

type CXXExpression = String

type CXXCode = Writer LinesBuildup

cxxLine :: String -> CXXCode ()

cxxIndent :: Int -> CXXCode a -> CXXCode a

cxxCodeString :: CXXCode () -> String

exportCXXinliningHeader :: FilePath -> CXXCode () -> IO ()