CodeGen.CXX.Cqtx.PhqFn

Description

While the cqtx library provides quite versatile and in principle very low-requirement fitting of general physical functions (namely without the need for any gradients etc.), in practise this is hampered by the cumbersomeness of defining the fittable_phmsqfn subclasses. The macros in fitfnmacros.h can only partially alleviate this: due to the limits of the C preprocessor, most of the calculation load needs to be done in also cumbersome C++ templates or, more likely, at runtime. At the moment, this is implemented rather poorly, in particular the almost bogosort-quality dimensional analysis.

This Haskell module does not claim to be an optimal solution, but it produces significantly better code than the CPP macros, with very safe and easy invocation. In the future this might become a lot more powerful, since it is also possible to automatically create highly optimised specialised versions of the functions, e.g. squaredistance-calculation on CUDA.

Documentation

phqFn

Arguments

:: forall paramsList . IsolenList paramsList

=> String
Name of the resulting phqFn C++ object

-> paramsList String
Default labels of the parameters

-> (forall x. PhqfnDefining x => paramsList x -> x)
Function definition, as a lambda

-> CqtxCode ()
C++ class and object code for a cqtx fittable physical function corresponding to the given definition.

Create a physical function that can be fitted to measured data using the cqtx algorithms such as evolution_minimizer. The invocation is similar to the primitive CPP macros, but type-safe and allows full Haskell syntax in the definition – though this can of course be exploited only so much, since phqfns are very limited in their abilities. To make sure these limits are maintained we use universally-quantised arguments (which also suits the implementation very well).

For example, the standard gaussian peak 𝐴⋅exp(-(𝑥-𝑥₀)²/(2⋅𝜎²)) could be defined thus:

  phqFn "gaussPeak" ("x":."x_0":."\\sigma":."A":.P)
                 (\ ( x :.  x0 :.  sigma  :. a :.P)
                   -> let q = (x-x0)/sigma in  a * exp(-0.5 * q^2) )

The use of the type-determined–length lists makes it impossible to accidentally give different numbers of parameter bindings and -labels.

Avoidance of duplicate calculation, as well as rudimentary optimisation, is taken care for by this preprocessor.

General Cqtx code generation

module CodeGen.CXX.Code

type CqtxCode = ReaderT CqtxConfig CXXCode

type CqtxConfig = ()

withDefaultCqtxConfig :: CqtxCode a -> CXXCode a

class Floating a => PhqfnDefining a where

Methods

sumOverIdx :: PhqVarIndexer -> ((IdxablePhqDefVar a -> a) -> a) -> a

Length-indexed lists

data IsolenEnd a

Constructors

P

data IsolenCons l a

Constructors

a :. (l a)