CodeGen.CXX.Code

Description

A writer monad for building up C++ code, to use Haskell as a (way more powerful) replacement for the C preprocessor. Unlike that, it is designed to produce reasonably human-readable code, so the resulting files can also be used as "proper" code files, without Haskell available.

Documentation

type CXXExpression = String

type CXXCode = Writer LinesBuildup

cxxLine :: String -> CXXCode ()

cxxIndent :: Int -> CXXCode a -> CXXCode a

cxxPreFirstLine :: CXXExpression -> CXXCode a -> CXXCode a

Add something in front of the first line, and indent the remaining ones accordingly so the relative alignment is preserved.

cxxPostLastLine :: CXXExpression -> CXXCode a -> CXXCode a

cxxSurround :: CXXExpression -> CXXCode a -> CXXExpression -> CXXCode a

procCXXCode :: (CXXExpression -> CXXExpression) -> CXXCode a -> CXXCode a

cxxConcatBlockIndented :: CXXCode a -> CXXCode b -> CXXCode b

cxxCombineBlockIndented :: (CXXExpression -> CXXExpression -> CXXExpression) -> CXXCode a -> CXXCode b -> CXXCode b

noCXXCode :: CXXCode ()

cxxCodeString :: CXXCode () -> String

makeSafeCXXIdentifier :: CXXExpression -> CXXExpression

exportCXXinliningHeader :: FilePath -> CXXCode () -> IO ()